refac: rename package name
