+++ src/jjjexperiment/v_min_input/section4_2.py
