+++ src/tests/v_min_input/test_solve_cubic_system.py
@@ -30,7 +30,7 @@ company_f = [
 ]
 
 import pytest
-from jjjexperiment.ac_min_volume_input.section4_2_a import _solve_cubic_system
+from jjjexperiment.v_min_input.section4_2_a import _solve_cubic_system
 
 class Test三乗方程式解法:
     """_solve_cubic_system関数のテスト"""
